# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.APIResponse:
    description: Standard API response format with metadata and context information
    properties:
      data:
        description: Response data (varies by endpoint)
      error:
        allOf:
        - $ref: '#/definitions/dto.ErrorResponse'
        description: Error details (only present when success=false)
      meta:
        allOf:
        - $ref: '#/definitions/dto.ResponseMetadata'
        description: Response metadata with request tracking
      pagination:
        allOf:
        - $ref: '#/definitions/dto.PaginationMetadata'
        description: Pagination metadata (for paginated responses)
      success:
        description: Indicates if the request was successful
        example: true
        type: boolean
    type: object
...
  dto.ErrorResponse:
    description: Error response structure with detailed error information
    properties:
      code:
        description: Error code for programmatic handling
        example: VALIDATION_ERROR
        type: string
      details:
        additionalProperties: true
        description: Additional error details
        type: object
      message:
        description: Human-readable error message
        example: Invalid input data
        type: string
      request_id:
        description: Request ID for tracking
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      timestamp:
        description: Error timestamp
        example: "2025-07-24T20:43:16.123456789+07:00"
        type: string
    type: object
...
  dto.PaginationMetadata:
    description: Pagination metadata for paginated responses
    properties:
      has_next:
        description: Whether there is a next page
        example: true
        type: boolean
      has_prev:
        description: Whether there is a previous page
        example: false
        type: boolean
      limit:
        description: Number of items per page
        example: 10
        type: integer
      page:
        description: Current page number (1-based)
        example: 1
        type: integer
      request_id:
        description: Request ID for tracking
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      total:
        description: Total number of items
        example: 100
        type: integer
      total_pages:
        description: Total number of pages
        example: 10
        type: integer
    type: object
  dto.ResponseMetadata:
    description: Response metadata containing request tracking and performance information
    properties:
      processing_time_ms:
        description: Request processing time in nanoseconds
        example: 15000000
        type: integer
      request_id:
        description: Unique request identifier for tracking
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      timestamp:
        description: Response generation timestamp
        example: "2025-07-24T20:43:16.123456789+07:00"
        type: string
      version:
        description: API version
        example: 1.0.0
        type: string
    type: object
...
  dto.UpdateCategoryRequest:
    properties:
      name:
        type: string
    type: object
...
  model.Article:
    properties:
      category:
        $ref: '#/definitions/model.Category'
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      slug:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/model.User'
      views:
        type: integer
    type: object
...
  model.Category:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  model.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:4300
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: REST API untuk aplikasi blog Develapar dengan fitur lengkap untuk manajemen
    artikel, komentar, kategori, tag, bookmark, dan like. API menggunakan standard
    response format dengan metadata, request tracking, rate limiting, dan comprehensive
    error handling.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Develapar API
  version: "1.0"
paths:
...
  /articles/category/{category_name}:
    get:
      description: Get a list of articles by category name
      parameters:
      - description: Name of the category to retrieve articles from
        in: path
        name: category_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of articles by category
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  properties:
                    articles:
                      items:
                        $ref: '#/definitions/model.Article'
                      type: array
                    message:
                      type: string
                  type: object
              type: object
        "400":
          description: Invalid category name
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "408":
          description: Request timeout
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
      summary: Get articles by category name
      tags:
      - Articles
...
  /categories:
    get:
      description: Get a list of all categories
      produces:
      - application/json
      responses:
        "200":
          description: List of categories
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  properties:
                    categories:
                      items:
                        $ref: '#/definitions/model.Category'
                      type: array
                    message:
                      type: string
                  type: object
              type: object
        "408":
          description: Request timeout
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
      summary: Get all categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Create a new category with a given name
      parameters:
      - description: Category creation details
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.Category'
      produces:
      - application/json
      responses:
        "201":
          description: Category successfully created
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  properties:
                    category:
                      $ref: '#/definitions/model.Category'
                    message:
                      type: string
                  type: object
              type: object
        "400":
          description: Invalid payload
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "408":
          description: Request timeout
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Create a new category
      tags:
      - Categories
  /categories/{category_id}:
    delete:
      description: Delete a category by ID
      parameters:
      - description: ID of the category to delete
        in: path
        name: category_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Category deleted successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  properties:
                    message:
                      type: string
                  type: object
              type: object
        "400":
          description: Invalid category ID
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "408":
          description: Request timeout
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Delete a category
      tags:
      - Categories
    get:
      description: Get category details by its ID
      parameters:
      - description: ID of the category to retrieve
        in: path
        name: category_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Category details
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  properties:
                    category:
                      $ref: '#/definitions/model.Category'
                    message:
                      type: string
                  type: object
              type: object
        "400":
          description: Invalid category ID
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "404":
          description: Category not found
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "408":
          description: Request timeout
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
      summary: Get category by ID
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: Update an existing category by ID
      parameters:
      - description: ID of the category to update
        in: path
        name: category_id
        required: true
        type: integer
      - description: Category update details
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Category updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  properties:
                    category:
                      $ref: '#/definitions/model.Category'
                    message:
                      type: string
                  type: object
              type: object
        "400":
          description: Invalid category ID or payload
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "408":
          description: Request timeout
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Update a category
      tags:
      - Categories
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"